Refactor migration files and adjust data types
Updated the migration files and data types to improve the application's performance. The change includes updating file names, changing the data type of PsuRating from int to short and removing unused lines. Additionally, the CatalogService and CatalogController classes were created to handle catalog data business logic and HTTP requests respectively.

+++ ThesisCatalog.API/Data/Migrations/ThesisCatalogDbContextModelSnapshot.cs
@@ -170,8 +170,8 @@ namespace ThesisCatalog.API.Data.Migrations
                             b1.Property<int>("UsbType")
                                 .HasColumnType("int");
 
-                            b1.Property<int>("PortCount")
-                                .HasColumnType("int");
+                            b1.Property<short>("PortCount")
+                                .HasColumnType("smallint");
 
                             b1.HasKey("CatalogItemId", "UsbType");
